devex-gpt-api/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==1.0.0
requests==2.26.0
flask-cors==3.0.10
